bug(tests): Remove PDF-related testing, update setup, refactor test-synonym_augmentation

+++ tests/test_augment_with_synonyms.py
@@ -9,10 +9,8 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 class TestAugmentation(unittest.TestCase):
     def setUp(self):
-        self.pdf_path = 'data/data_aumentation_in_natural_language_processing.pdf'
         self.model_response_path = 'data/model_response.txt'
-        self.augmented_response_path = 'data/augmented_response.txt'
-        self._check_file_exists(self.pdf_path, "PDF")
+        self.augmented_response_path = 'data/synonym_augmented_model_response.txt'
         self._check_file_exists(self.model_response_path, "Model response")
 
     def _check_file_exists(self, file_path, file_description):
@@ -20,13 +18,9 @@ class TestAugmentation(unittest.TestCase):
             raise FileNotFoundError(f"The {file_description} file '{
                                     file_path}' does not exist")
 
-    def test_pdf_exists(self):
-        self.assertTrue(os.path.exists(self.pdf_path), f"The PDF file '{
-                        self.pdf_path}' does not exist")
-
     def test_model_response_text_exists(self):
-        self.assertTrue(os.path.exists(self.model_response_path), f"The Model response file '{
-                        self.model_response_path}' does not exist")
+        self.assertTrue(os.path.exists(self.model_response_path),
+                        f"The Model response file '{self.model_response_path}' does not exist")
 
     def test_synonym_augmentation(self):
         # Perform synonym augmentation
@@ -34,8 +28,8 @@ class TestAugmentation(unittest.TestCase):
                               self.augmented_response_path)
 
         # Check if the augmented file was created
-        self.assertTrue(os.path.exists(self.augmented_response_path), f"The augmented response file '{
-                        self.augmented_response_path}' was not created")
+        self.assertTrue(os.path.exists(self.augmented_response_path),
+                        f"The augmented response file '{self.augmented_response_path}' was not created")
 
         # Check if the augmented file is different from the original
         with open(self.model_response_path, 'r') as original, open(self.augmented_response_path, 'r') as augmented:
@@ -44,8 +38,10 @@ class TestAugmentation(unittest.TestCase):
             self.assertNotEqual(original_content, augmented_content,
                                 "The augmented content should be different from the original")
 
+    def tearDown(self):
         # Clean up the augmented file after the test
-        os.remove(self.augmented_response_path)
+        if os.path.exists(self.augmented_response_path):
+            os.remove(self.augmented_response_path)
 
 
 if __name__ == '__main__':
